requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# python-multipart==0.0.6
# sqlalchemy==2.0.23
# pydantic==2.4.2
# python-dotenv==1.0.0
# psycopg2-binary==2.9.9
# alembic==1.12.1
# python-jose==3.3.0
# email-validator==2.1.0
# python-socketio==5.9.0
# python-dateutil==2.8.2
# pydantic-settings==2.0.3
# python-multipart==0.0.6
# httpx==0.25.0


fastapi==0.116.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.6
sqlalchemy==2.0.43
pydantic==2.11.7
python-dotenv==1.1.1
alembic==1.12.1
email-validator==2.1.0
python-socketio==5.9.0
python-dateutil==2.8.2
pydantic-settings==2.0.3
httpx==0.25.0
greenlet==3.2.4
asyncpg==0.29.0
aiosqlite==0.19.0
sqlalchemy[asyncio]==2.0.43